refactor(inertia client-side config): move config to /lib and centers whole page resolver function in the shared constants
