+++ www/app.tsx
@@ -5,7 +5,7 @@ import { createInertiaApp } from "@inertiajs/react";
 import { createRoot, hydrateRoot } from "react-dom/client";
 
 import { appConfig } from "./config/app";
-import { type PageComponent, resolvePageLayout, resolveTitle } from "./inertiaShared";
+import { pageResolver, resolveTitle } from "./lib/inertia";
 
 const appName = appConfig.appName;
 
@@ -14,13 +14,7 @@ createInertiaApp({
 
   title: (title) => resolveTitle(title, appName),
 
-  resolve: async (name) => {
-    const pages = import.meta.glob("./pages/**/*.tsx", { eager: true });
-    const page = pages[`./pages/${name}.tsx`] as PageComponent;
-    const resolvedPage = resolvePageLayout(page);
-
-    return resolvedPage;
-  },
+  resolve: pageResolver,
 
   setup({ el, App, props }) {
     const isSSR = document.head
